Start cleanup of WorldEdit-CLI data versions
