+++ worldedit-cli/src/main/java/com/sk89q/worldedit/cli/CLIBlockRegistry.java
@@ -21,7 +21,7 @@ package com.sk89q.worldedit.cli;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
-import com.sk89q.worldedit.cli.data.FileRegistries;
+import com.sk89q.worldedit.cli.data.DataFile;
 import com.sk89q.worldedit.registry.state.BooleanProperty;
 import com.sk89q.worldedit.registry.state.DirectionalProperty;
 import com.sk89q.worldedit.registry.state.EnumProperty;
@@ -51,9 +51,9 @@ public class CLIBlockRegistry extends BundledBlockRegistry {
     @Nullable
     @Override
     public Map<String, ? extends Property<?>> getProperties(BlockType blockType) {
-        Map<String, FileRegistries.BlockProperty> properties =
+        Map<String, DataFile.BlockProperty> properties =
                 CLIWorldEdit.inst.getFileRegistries().getDataFile().blocks.get(blockType.id()).properties;
-        Maps.EntryTransformer<String, FileRegistries.BlockProperty, Property<?>> entryTransform =
+        Maps.EntryTransformer<String, DataFile.BlockProperty, Property<?>> entryTransform =
             (key, value) -> createProperty(value.type, key, value.values);
         return ImmutableMap.copyOf(Maps.transformEntries(properties, entryTransform));
     }
